Split session/mod.rs by phase; fix audio-spawn teardown race
session/ was one 1156-line file mixing eight unrelated concerns. Split
into connect (session bring-up), pump (video/audio thread bodies),
probe (pairing trust step and speed test handshakes), and the shared
sink, timeline, stats, priority, join. mod.rs is now docs + re-exports.

session::connect took 15 positional args behind an
allow(clippy::too_many_arguments); now a ConnectParams struct with the
body split into five named helpers (capability clamp, handshake dial,
handshake logging, decoder load, audio-path labelling).

The video pump loop became a VideoPump struct (on_frame/heartbeat/
note_loss/forward_hdr_meta); its ad-hoc Instant deadlines became a
Tick helper with named constants instead of bare 500ms/2s/15s
literals. The two copy-pasted audio pump threads (NDL-offloaded and
software-decoded) now share one audio_drain closure.

sink.rs's submit() threaded a mutable need_keyframe flag through a
hold/feed/error sequence; now a gate + feed + error handler with a
HoldGate enum making "fed but still want a keyframe" explicit.
probe.rs's two 18-argument NativeClient::connect calls now share one
handshake_only helper so the shared invariants (throwaway 720p, whole
access units, no cursor, required ChaCha20) are stated once.

session/ no longer imports sdl2: the gamepad feedback drain moved to
runtime/session_ext.rs, which owns the SDL objects; the renice syscall
wrapper and /proc self readers moved to platform/webos/device.rs,
removing evdev.rs's duplicate raw setpriority call.

Fixes a real bug found during the split: if the audio thread failed
to spawn after the video thread was already running, the error path
returned without setting the session stop flag, leaving the video
pump detached and feeding NDL while the caller could go on to call
ndl::quit(), racing a live pump against an unload. It now stops and
joins the video thread before returning the error.

Behavior otherwise unchanged. request_access and run_speed_probe move
to session::probe::* for in-crate callers. CLAUDE.md's architecture
section is updated to match (it referenced a session/pacing.rs that
no longer exists).
